project completion

+++ app/Http/Controllers/PostsController.php
@@ -4,9 +4,21 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Post;
+use Illuminate\Support\Facades\Storage;
 
 class PostsController extends Controller
 {
+
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except'=>['index','show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -38,13 +50,27 @@ class PostsController extends Controller
     {
         $this->validate($request,[
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        //handle cover_image
+        if ($request->hasFile('cover_image')) {
+            //get file name
+            $fileNameToStore = md5(time()).'.'.$request->file('cover_image')->getClientOriginalExtension();
+            $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }
+        else
+        {
+            $fileNameToStore = 'noimage.jpg';
+        }
 
         //create post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success','Post Created');
@@ -72,6 +98,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error','No edit permission');
+        }
+        
         return view('posts.edit')->with('post', $post);
     }
 
@@ -86,13 +117,28 @@ class PostsController extends Controller
     {
         $this->validate($request,[
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
 
-        //create post
+        //handle cover_image
+        if ($request->hasFile('cover_image')) {
+            //get file name
+            $fileNameToStore = md5(time()).'.'.$request->file('cover_image')->getClientOriginalExtension();
+            $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }
+        
+
+        //update post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if ($request->hasFile('cover_image')) {
+            //remove orginal
+            Storage::delete('public/cover_images/'.$post->cover_image);
+            //update to new image
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success','Post Updated');
@@ -107,6 +153,14 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        if (auth()->user()->id != $post->user_id) {
+            return redirect('/posts')->with('error','No delete permission');
+        }
+        if ($post->cover_image != 'noimage.jpg') {
+            //delete image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
         $post->delete();
         return redirect('/posts')->with('success','Post Removed');
     }
